+++ manim/renderer/opengl_renderer.py
@@ -1,21 +1,16 @@
-import inspect
 import itertools as it
-import threading
 import time
 from queue import Queue
 
 import moderngl
 import numpy as np
 from PIL import Image
-from watchdog.events import FileSystemEventHandler
-from watchdog.observers import Observer
 
 from manim import config, logger
 from manim.utils.color import color_to_rgba
 from manim.utils.exceptions import EndSceneEarlyException, RerunSceneException
 
 from ..constants import *
-from ..gui.gui import configure_pygui
 from ..mobject.opengl_mobject import OpenGLMobject, OpenGLPoint
 from ..mobject.types.opengl_vectorized_mobject import OpenGLVMobject
 from ..utils import opengl, space_ops
@@ -34,24 +29,6 @@ from .vectorized_mobject_rendering import (
     render_opengl_vectorized_mobject_stroke,
 )
 
-try:
-    import dearpygui.dearpygui as dpg
-
-    dearpygui_imported = True
-except ImportError:
-    dearpygui_imported = False
-
-
-class RerunSceneHandler(FileSystemEventHandler):
-    """A class to handle rerunning a Scene after the input file is modified."""
-
-    def __init__(self, queue):
-        super().__init__()
-        self.queue = queue
-
-    def on_modified(self, event):
-        self.queue.put(("rerun_file", [], {}))
-
 
 class OpenGLCamera(OpenGLMobject):
     euler_angles = _Data()
@@ -68,8 +45,10 @@ class OpenGLCamera(OpenGLMobject):
         minimum_polar_angle=-PI / 2,
         maximum_polar_angle=PI / 2,
         model_matrix=None,
+            scene=None,
         **kwargs,
     ):
+        self.scene = scene # todo remove
         self.use_z_index = True
         self.frame_rate = 60
         self.orthographic = orthographic
@@ -252,26 +231,22 @@ class OpenGLRenderer:
         self.animations_hashes = []
         self.num_plays = 0
         self.interactive_mode = False
-        self.queue = Queue()
         self.widgets = [] if not widgets else widgets
         self.key_to_function_map = {}
         self.mouse_press_callbacks = []
 
         self.camera = OpenGLCamera()
-        self.pressed_keys = set()
 
         # Initialize texture map.
         self.path_to_texture_id = {}
 
         self._background_color = color_to_rgba(config["background_color"], 1.0)
 
-    def init_scene(self):
+    def init_scene(self, scene=None):
+        self.scene = scene # todo remove
         if not hasattr(self, "window"):
             if self.should_create_window():
                 from .opengl_renderer_window import Window
-
-                self.mouse_point = OpenGLPoint()
-                self.mouse_drag_point = OpenGLPoint()
 
                 self.window = Window(self)
                 self.context = self.window.ctx
@@ -559,340 +534,6 @@ class OpenGLRenderer:
             # Only scale wrt one axis
             scale = fh / ph
             return fc + scale * np.array([(px - pw / 2), (py - ph / 2), 0])
-
-    def interactive_embed(
-        self,
-        moving_mobjects,
-        meshes,
-        file_writer,
-        update_mobjects_hook,
-        update_meshes_hook,
-        update_scene_hook,
-    ):
-        """
-        Like embed(), but allows for screen interaction.
-        """
-        if not self.check_interactive_embed_is_valid():
-            return
-        self.interactive_mode = True
-
-        def ipython(shell, namespace):
-            import manim.opengl
-
-            def load_module_into_namespace(module, namespace):
-                for name in dir(module):
-                    namespace[name] = getattr(module, name)
-
-            load_module_into_namespace(manim, namespace)
-            load_module_into_namespace(manim.opengl, namespace)
-
-            def embedded_rerun(*args, **kwargs):
-                self.queue.put(("rerun_keyboard", args, kwargs))
-                shell.exiter()
-
-            namespace["rerun"] = embedded_rerun
-
-            shell(local_ns=namespace)
-            self.queue.put(("exit_keyboard", [], {}))
-
-        def get_embedded_method(method_name):
-            return lambda *args, **kwargs: self.queue.put((method_name, args, kwargs))
-
-        local_namespace = inspect.currentframe().f_back.f_locals
-        for method in ("play", "wait", "add", "remove"):
-            embedded_method = get_embedded_method(method)
-            # Allow for calling scene methods without prepending 'self.'.
-            local_namespace[method] = embedded_method
-
-        from IPython.terminal.embed import InteractiveShellEmbed
-        from traitlets.config import Config
-
-        cfg = Config()
-        cfg.TerminalInteractiveShell.confirm_exit = False
-        shell = InteractiveShellEmbed(config=cfg)
-
-        keyboard_thread = threading.Thread(
-            target=ipython,
-            args=(shell, local_namespace),
-        )
-        # run as daemon to kill thread when main thread exits
-        if not shell.pt_app:
-            keyboard_thread.daemon = True
-        keyboard_thread.start()
-
-        if dearpygui_imported and config["enable_gui"]:
-            if not dpg.is_dearpygui_running():
-                gui_thread = threading.Thread(
-                    target=configure_pygui,
-                    args=(self, self.widgets),
-                    kwargs={"update": False},
-                )
-                gui_thread.start()
-            else:
-                configure_pygui(self, self.widgets, update=True)
-
-        self.camera.model_matrix = self.camera.default_model_matrix
-
-        self.interact(
-            shell,
-            keyboard_thread,
-            moving_mobjects,
-            meshes,
-            file_writer,
-            update_mobjects_hook,
-            update_meshes_hook,
-            update_scene_hook,
-        )
-
-    def interact(
-        self,
-        shell,
-        keyboard_thread,
-        moving_mobjects,
-        meshes,
-        file_writer,
-        update_mobjects_hook,
-        update_meshes_hook,
-        update_scene_hook,
-    ):
-        event_handler = RerunSceneHandler(self.queue)
-        file_observer = Observer()
-        file_observer.schedule(event_handler, config["input_file"], recursive=True)
-        file_observer.start()
-
-        self.quit_interaction = False
-        keyboard_thread_needs_join = shell.pt_app is not None
-        assert self.queue.qsize() == 0
-
-        last_time = time.time()
-        while not (self.window.is_closing or self.quit_interaction):
-            if not self.queue.empty():
-                tup = self.queue.get_nowait()
-                if tup[0].startswith("rerun"):
-                    # Intentionally skip calling join() on the file thread to save time.
-                    if not tup[0].endswith("keyboard"):
-                        if shell.pt_app:
-                            shell.pt_app.app.exit(exception=EOFError)
-                        file_observer.unschedule_all()
-                        raise RerunSceneException
-                    keyboard_thread.join()
-
-                    kwargs = tup[2]
-                    if "from_animation_number" in kwargs:
-                        config["from_animation_number"] = kwargs[
-                            "from_animation_number"
-                        ]
-                    # # TODO: This option only makes sense if interactive_embed() is run at the
-                    # # end of a scene by default.
-                    # if "upto_animation_number" in kwargs:
-                    #     config["upto_animation_number"] = kwargs[
-                    #         "upto_animation_number"
-                    #     ]
-
-                    keyboard_thread.join()
-                    file_observer.unschedule_all()
-                    raise RerunSceneException
-                elif tup[0].startswith("exit"):
-                    # Intentionally skip calling join() on the file thread to save time.
-                    if not tup[0].endswith("keyboard") and shell.pt_app:
-                        shell.pt_app.app.exit(exception=EOFError)
-                    keyboard_thread.join()
-                    # Remove exit_keyboard from the queue if necessary.
-                    while self.queue.qsize() > 0:
-                        self.queue.get()
-                    keyboard_thread_needs_join = False
-                    break
-                else:
-                    method, args, kwargs = tup
-                    getattr(self, method)(*args, **kwargs)
-            else:
-                self.animation_start_time = 0
-                dt = time.time() - last_time
-                last_time = time.time()
-                self.render(
-                    dt,
-                    moving_mobjects,
-                    mobjects=moving_mobjects,
-                    meshes=meshes,
-                    file_writer=file_writer,
-                )
-                update_mobjects_hook(dt)
-                update_meshes_hook(dt)
-                update_scene_hook(dt)
-
-        # Join the keyboard thread if necessary.
-        if shell is not None and keyboard_thread_needs_join:
-            shell.pt_app.app.exit(exception=EOFError)
-            keyboard_thread.join()
-            # Remove exit_keyboard from the queue if necessary.
-            while self.queue.qsize() > 0:
-                self.queue.get()
-
-        file_observer.stop()
-        file_observer.join()
-
-        if dearpygui_imported and config["enable_gui"]:
-            dpg.stop_dearpygui()
-
-        if self.window.is_closing:
-            self.window.destroy()
-
-    def check_interactive_embed_is_valid(self):
-        if config["force_window"]:
-            return True
-        if config["write_to_movie"]:
-            logger.warning("Disabling interactive embed as 'write_to_movie' is enabled")
-            return False
-        elif config["format"]:
-            logger.warning(
-                "Disabling interactive embed as '--format' is set as "
-                + config["format"],
-            )
-            return False
-        elif not self.window:
-            logger.warning("Disabling interactive embed as no window was created")
-            return False
-        elif config.dry_run:
-            logger.warning("Disabling interactive embed as dry_run is enabled")
-            return False
-        return True
-
-    def on_mouse_motion(self, point, d_point):
-        self.mouse_point.move_to(point)
-        if SHIFT_VALUE in self.pressed_keys:
-            shift = -d_point
-            shift[0] *= self.camera.get_width() / 2
-            shift[1] *= self.camera.get_height() / 2
-            transform = self.camera.inverse_rotation_matrix
-            shift = np.dot(np.transpose(transform), shift)
-            self.camera.shift(shift)
-
-    def on_mouse_scroll(self, point, offset):
-        if not config.use_projection_stroke_shaders:
-            factor = 1 + np.arctan(-2.1 * offset[1])
-            self.camera.scale(factor, about_point=self.camera_target)
-        self.mouse_scroll_orbit_controls(point, offset)
-
-    def on_key_press(self, symbol, modifiers):
-        try:
-            char = chr(symbol)
-        except OverflowError:
-            logger.warning("The value of the pressed key is too large.")
-            return
-
-        if char == "r":
-            self.camera.to_default_state()
-            self.camera_target = np.array([0, 0, 0], dtype=np.float32)
-        elif char == "q":
-            self.quit_interaction = True
-        else:
-            if char in self.key_to_function_map:
-                self.key_to_function_map[char]()
-
-    def on_key_release(self, symbol, modifiers):
-        pass
-
-    def on_mouse_drag(self, point, d_point, buttons, modifiers):
-        self.mouse_drag_point.move_to(point)
-        if buttons == 1:
-            self.camera.increment_theta(-d_point[0])
-            self.camera.increment_phi(d_point[1])
-        elif buttons == 4:
-            camera_x_axis = self.camera.model_matrix[:3, 0]
-            horizontal_shift_vector = -d_point[0] * camera_x_axis
-            vertical_shift_vector = -d_point[1] * np.cross(OUT, camera_x_axis)
-            total_shift_vector = horizontal_shift_vector + vertical_shift_vector
-            self.camera.shift(1.1 * total_shift_vector)
-
-        self.mouse_drag_orbit_controls(point, d_point, buttons, modifiers)
-
-    def mouse_scroll_orbit_controls(self, point, offset):
-        camera_to_target = self.camera_target - self.camera.get_position()
-        camera_to_target *= np.sign(offset[1])
-        shift_vector = 0.01 * camera_to_target
-        self.camera.model_matrix = (
-            opengl.translation_matrix(*shift_vector) @ self.camera.model_matrix
-        )
-
-    def mouse_drag_orbit_controls(self, point, d_point, buttons, modifiers):
-        # Left click drag.
-        if buttons == 1:
-            # Translate to target the origin and rotate around the z axis.
-            self.camera.model_matrix = (
-                opengl.rotation_matrix(z=-d_point[0])
-                @ opengl.translation_matrix(*-self.camera_target)
-                @ self.camera.model_matrix
-            )
-
-            # Rotation off of the z axis.
-            camera_position = self.camera.get_position()
-            camera_y_axis = self.camera.model_matrix[:3, 1]
-            axis_of_rotation = space_ops.normalize(
-                np.cross(camera_y_axis, camera_position),
-            )
-            rotation_matrix = space_ops.rotation_matrix(
-                d_point[1],
-                axis_of_rotation,
-                homogeneous=True,
-            )
-
-            maximum_polar_angle = self.camera.maximum_polar_angle
-            minimum_polar_angle = self.camera.minimum_polar_angle
-
-            potential_camera_model_matrix = rotation_matrix @ self.camera.model_matrix
-            potential_camera_location = potential_camera_model_matrix[:3, 3]
-            potential_camera_y_axis = potential_camera_model_matrix[:3, 1]
-            sign = (
-                np.sign(potential_camera_y_axis[2])
-                if potential_camera_y_axis[2] != 0
-                else 1
-            )
-            potential_polar_angle = sign * np.arccos(
-                potential_camera_location[2]
-                / np.linalg.norm(potential_camera_location),
-            )
-            if minimum_polar_angle <= potential_polar_angle <= maximum_polar_angle:
-                self.camera.model_matrix = potential_camera_model_matrix
-            else:
-                sign = np.sign(camera_y_axis[2]) if camera_y_axis[2] != 0 else 1
-                current_polar_angle = sign * np.arccos(
-                    camera_position[2] / np.linalg.norm(camera_position),
-                )
-                if potential_polar_angle > maximum_polar_angle:
-                    polar_angle_delta = maximum_polar_angle - current_polar_angle
-                else:
-                    polar_angle_delta = minimum_polar_angle - current_polar_angle
-                rotation_matrix = space_ops.rotation_matrix(
-                    polar_angle_delta,
-                    axis_of_rotation,
-                    homogeneous=True,
-                )
-                self.camera.model_matrix = rotation_matrix @ self.camera.model_matrix
-
-            # Translate to target the original target.
-            self.camera.model_matrix = (
-                opengl.translation_matrix(*self.camera_target)
-                @ self.camera.model_matrix
-            )
-        # Right click drag.
-        elif buttons == 4:
-            camera_x_axis = self.camera.model_matrix[:3, 0]
-            horizontal_shift_vector = -d_point[0] * camera_x_axis
-            vertical_shift_vector = -d_point[1] * np.cross(OUT, camera_x_axis)
-            total_shift_vector = horizontal_shift_vector + vertical_shift_vector
-
-            self.camera.model_matrix = (
-                opengl.translation_matrix(*total_shift_vector)
-                @ self.camera.model_matrix
-            )
-            self.camera_target += total_shift_vector
-
-    def set_key_function(self, char, func):
-        self.key_to_function_map[char] = func
-
-    def on_mouse_press(self, point, button, modifiers):
-        for func in self.mouse_press_callbacks:
-            func()
 
     @property
     def background_color(self):
