+++ manim/renderer/opengl_renderer_window.py
@@ -55,7 +55,7 @@ class Window(PygletWindow):
         super().on_mouse_motion(x, y, dx, dy)
         point = self.renderer.pixel_coords_to_space_coords(x, y)
         d_point = self.renderer.pixel_coords_to_space_coords(dx, dy, relative=True)
-        self.renderer.on_mouse_motion(point, d_point)
+        self.renderer.scene.on_mouse_motion(point, d_point)
 
     def on_mouse_scroll(self, x, y, x_offset: float, y_offset: float):
         super().on_mouse_scroll(x, y, x_offset, y_offset)
@@ -65,24 +65,24 @@ class Window(PygletWindow):
             y_offset,
             relative=True,
         )
-        self.renderer.on_mouse_scroll(point, offset)
+        self.renderer.scene.on_mouse_scroll(point, offset)
 
     def on_key_press(self, symbol, modifiers):
         self.renderer.pressed_keys.add(symbol)
         super().on_key_press(symbol, modifiers)
-        self.renderer.on_key_press(symbol, modifiers)
+        self.renderer.scene.on_key_press(symbol, modifiers)
 
     def on_key_release(self, symbol, modifiers):
         if symbol in self.renderer.pressed_keys:
             self.renderer.pressed_keys.remove(symbol)
         super().on_key_release(symbol, modifiers)
-        self.renderer.on_key_release(symbol, modifiers)
+        self.renderer.scene.on_key_release(symbol, modifiers)
 
     def on_mouse_drag(self, x, y, dx, dy, buttons, modifiers):
         super().on_mouse_drag(x, y, dx, dy, buttons, modifiers)
         point = self.renderer.pixel_coords_to_space_coords(x, y)
         d_point = self.renderer.pixel_coords_to_space_coords(dx, dy, relative=True)
-        self.renderer.on_mouse_drag(point, d_point, buttons, modifiers)
+        self.renderer.scene.on_mouse_drag(point, d_point, buttons, modifiers)
 
     def find_initial_position(self, size, monitor):
         custom_position = config.window_position
@@ -122,4 +122,4 @@ class Window(PygletWindow):
             2: "MOUSE",
             4: "RIGHT",
         }
-        self.renderer.on_mouse_press(point, mouse_button_map[button], modifiers)
+        self.renderer.scene.on_mouse_press(point, mouse_button_map[button], modifiers)
